Some good works in regards to projectiles

+++ MagusEngine/Systems/Time/Nodes/ComponentTimeNode.cs
@@ -6,9 +6,9 @@ namespace MagusEngine.Systems.Time.Nodes
     {
         public long Tick { get; }
         public uint Id { get; }
-        public Func<uint> Action { get; }
+        public Func<long> Action { get; }
 
-        public ComponentTimeNode(long tick, uint id, Func<uint> action)
+        public ComponentTimeNode(long tick, uint id, Func<long> action)
         {
             Tick = tick;
             Id = id;
